Add default User image

+++ backend/src/routes/recipe.routes.test.ts
@@ -233,7 +233,7 @@ const removeDummyFiles = () => {
         files.forEach(file => {
             const fileDir = path.join(imagePath, file);
             if (file !== 'test_burger.jpeg' && file !== 'test_falafel.jpeg' && file !== 'test_spaghetti.jpeg'
-                && file !== 'profile-dummy.jpg' && file !== 'recipe-dummy.png') {
+                && file !== 'profile-dummy.jpg' && file !== 'recipe-dummy.png' && file !== 'User1.png') {
                 fs.unlinkSync(fileDir);
             }
         });
